requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
argcomplete==3.6.2
asgi-csrf==0.11
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.40.14
botocore==1.40.14
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-default-group==1.2.4
cohere==5.17.0
colorama==0.4.6
colorlog==6.9.0
datasette==0.65.1
distro==1.9.0
Django==5.2.5
djangorestframework==3.16.1
dnspython==2.7.0
drf-pydantic==2.9.0
email_validator==2.2.0
eval_type_backport==0.2.2
fake-useragent==2.2.0
Faker==37.5.3
fastavro==1.12.0
filelock==3.19.1
flexcache==0.3
flexparser==0.4
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.31.0
griffe==1.12.1
groq==0.31.0
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
hupper==1.12.1
idna==3.10
importlib_metadata==8.7.0
invoke==2.2.0
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
json-schema-to-pydantic==0.2.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
logfire-api==4.3.3
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mcp==1.13.0
mdurl==0.1.2
mergedeep==1.3.4
mistralai==1.9.7
multidict==6.6.4
nest-asyncio==1.6.0
openai==1.100.2
opentelemetry-api==1.36.0
packaging==25.0
peewee==3.18.2
platformdirs==4.3.8
pluggy==1.6.0
polyfactory==2.22.2
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-ai==0.0.43
pydantic-ai-slim==0.0.43
pydantic-graph==0.0.43
pydantic-settings==2.10.1
pydantic2==1.1.15
pydantic3==1.0.14
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
questionary==2.1.0
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
s3transfer==0.13.1
semver==3.0.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
tokenizers==0.21.4
tqdm==4.67.1
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
